# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample rk-demo echo server.",
        "title": "Demo Echo API Framework",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/buy/{wager_id}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wager"
                ],
                "summary": "Buy a wager",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the wager to buy",
                        "name": "wager_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Input Required",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wagerdto.BuyWagerInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errutil.EchoErrorStruct"
                        }
                    }
                }
            }
        },
        "/wagers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wager"
                ],
                "summary": "List all wager",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errutil.EchoErrorStruct"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wager"
                ],
                "summary": "Place a wager",
                "parameters": [
                    {
                        "description": "Input Required",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wagerdto.PlaceWagerInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Wager"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errutil.EchoErrorStruct"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errutil.EchoErrorStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "info": {}
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "bought_at": {
                    "type": "string"
                },
                "buying_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "wager_id": {
                    "type": "integer"
                }
            }
        },
        "models.Wager": {
            "type": "object",
            "properties": {
                "amount_sold": {
                    "type": "number"
                },
                "current_selling_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "odds": {
                    "type": "number"
                },
                "percentage_sold": {
                    "type": "number"
                },
                "placed_at": {
                    "type": "string"
                },
                "selling_percentage": {
                    "type": "number"
                },
                "selling_price": {
                    "type": "number"
                },
                "total_wager_value": {
                    "type": "number"
                }
            }
        },
        "wagerdto.BuyWagerInput": {
            "type": "object",
            "required": [
                "buying_price"
            ],
            "properties": {
                "buying_price": {
                    "type": "number"
                }
            }
        },
        "wagerdto.PlaceWagerInput": {
            "type": "object",
            "required": [
                "odds",
                "selling_percentage",
                "selling_price",
                "total_wager_value"
            ],
            "properties": {
                "odds": {
                    "type": "number"
                },
                "selling_percentage": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 1
                },
                "selling_price": {
                    "type": "number"
                },
                "total_wager_value": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}